extract SingleInputTestController, fix valueOrElse edge case
